Add textures for machine blocks
Added textures for new machine blocks, including items on display when smelting in a smeltery
Redesigned compressors to have the same texture on all sides and have an input pipe that covers the head.
Slightly tweaked Custom Crafter and Master Crafter textures to make them clearer and fit better with the rest of machine blocks.